рефакторинг слоя работы с данными: единый persistentContainer, уточнение протоколов
